# internal/store/db/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.27.0

package db

import (
	"github.com/jackc/pgx/v5/pgtype"
)

type Dataset struct {
	ID          string             `json:"id"`
	Name        string             `json:"name"`
	Description string             `json:"description"`
	CreatedAt   pgtype.Timestamptz `json:"created_at"`
}

type EventFile struct {
	ID        string             `json:"id"`
	DatasetID string             `json:"dataset_id"`
	Path      string             `json:"path"`
	Format    string             `json:"format"`
	Symbol    string             `json:"symbol"`
	Bytes     int64              `json:"bytes"`
	CreatedAt pgtype.Timestamptz `json:"created_at"`
	Sha256    string             `json:"sha256"`
	Rows      int64              `json:"rows"`
}

type ReplayJob struct {
	ID             string             `json:"id"`
	DatasetID      string             `json:"dataset_id"`
	EventFileID    string             `json:"event_file_id"`
	IdempotencyKey pgtype.Text        `json:"idempotency_key"`
	Symbol         string             `json:"symbol"`
	Speed          string             `json:"speed"`
	Status         string             `json:"status"`
	Attempts       int32              `json:"attempts"`
	LastError      string             `json:"last_error"`
	CheckpointLine int64              `json:"checkpoint_line"`
	CreatedAt      pgtype.Timestamptz `json:"created_at"`
	StartedAt      pgtype.Timestamptz `json:"started_at"`
	CompletedAt    pgtype.Timestamptz `json:"completed_at"`
	CanceledAt     pgtype.Timestamptz `json:"canceled_at"`
	Manifest       []byte             `json:"manifest"`
}

type ReplayMetric struct {
	ID              string             `json:"id"`
	JobID           string             `json:"job_id"`
	Rows            int64              `json:"rows"`
	Events          int64              `json:"events"`
	MalformedEvents int64              `json:"malformed_events"`
	SequenceGaps    int64              `json:"sequence_gaps"`
	DurationNs      int64              `json:"duration_ns"`
	RowsPerSecond   float64            `json:"rows_per_second"`
	EventsPerSecond float64            `json:"events_per_second"`
	P95LatencyNs    int64              `json:"p95_latency_ns"`
	PeakAllocBytes  int64              `json:"peak_alloc_bytes"`
	AllocsPerEvent  float64            `json:"allocs_per_event"`
	CreatedAt       pgtype.Timestamptz `json:"created_at"`
}

type ValidationError struct {
	ID        string             `json:"id"`
	JobID     string             `json:"job_id"`
	Line      int64              `json:"line"`
	Symbol    string             `json:"symbol"`
	Type      string             `json:"type"`
	Message   string             `json:"message"`
	CreatedAt pgtype.Timestamptz `json:"created_at"`
}